+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## 2026-05-27 - Git commit version stamp
 
+- Version: `0.1.8`
+- Timestamp: 2026-05-27 05:45 (UTC+7)
+- Commit: `a2907da`
+- Type: Patch
+- Status: Draft
+
+### Changes
+
+- Version stamp for git commit.
+
+### Verification
+
+- pending
+
+---
+## 2026-05-27 - Git commit version stamp
+
 - Version: `0.1.7`
 - Timestamp: 2026-05-27 04:41 (UTC+7)
 - Commit: `19f8789`